New `positron-data-explorer-duckdb` extension
