+++ extensions/positron-data-driver-duckdb/package.json
@@ -23,6 +23,7 @@
   },
   "devDependencies": {
     "@types/mocha": "^9.1.0",
+    "positron-data-explorer-duckdb": "file:../positron-data-explorer-duckdb",
     "positron-data-explorer-protocol": "file:../positron-data-explorer-protocol",
     "@types/node": "14.x",
     "@typescript-eslint/eslint-plugin": "^5.12.1",
