+++ extensions/positron-data-explorer-duckdb/src/duckdbWorkerClient.ts
