+++ extensions/positron-data-explorer-duckdb/src/duckdbWorkerPool.ts
