+++ extensions/positron-data-explorer-duckdb/src/duckdbWorkerProtocol.ts
