+++ extensions/positron-data-explorer-duckdb/src/workerEnv.ts
